third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/source-list-parsing-no-semicolon-expected.txt

CONSOLE ERROR: The Content-Security-Policy directive 'default-src' contains 'script-src' as a source expression. Did you want to add it as a directive and forget a semicolon?
CONSOLE ERROR: The Content-Security-Policy directive 'script-src' contains 'object-src' as a source expression. Did you want to add it as a directive and forget a semicolon?
CONSOLE ERROR: The Content-Security-Policy directive 'script-src' contains 'style-src' as a source expression. Did you want to add it as a directive and forget a semicolon?
If a web author forgets a semicolon, we should do our best to warn them that the policy they've defined is probably not what they intended.

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
PASS
